Plugged in logging ... cleaned up evaluators ... created base class for temporal types - remove dup code

+++ cql-engine/src/test/java/org/opencds/cqf/cql/execution/CqlClinicalOperatorsTest.java
@@ -38,19 +38,19 @@ public class CqlClinicalOperatorsTest extends CqlExecutionTestBase {
         // assertThat(result, is(16));
 
         Object result = context.resolveExpressionRef("CalculateAgeMonths").getExpression().evaluate(context);
-        assertThat(result, is((Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(3), new int[] {2000, 1, 1})), DateTime.getToday(), "month")));
+        assertThat(result, is((CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(3), new int[] {2000, 1, 1})), DateTime.getToday(), "month"))));
 
         result = context.resolveExpressionRef("CalculateAgeDays").getExpression().evaluate(context);
-        assertThat(result, is((Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(3), new int[] {2000, 1, 1})), DateTime.getToday(), "day")));
+        assertThat(result, is(CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(3), new int[] {2000, 1, 1})), DateTime.getToday(), "day")));
 
         result = context.resolveExpressionRef("CalculateAgeHours").getExpression().evaluate(context);
-        assertThat(result, is((Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(4), new int[] {2000, 1, 1, 0})), DateTime.getToday(), "hour")));
+        assertThat(result, is(CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(4), new int[] {2000, 1, 1, 0})), DateTime.getToday(), "hour")));
 
         result = context.resolveExpressionRef("CalculateAgeMinutes").getExpression().evaluate(context);
-        assertThat(result, is((Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(5), new int[] {2000, 1, 1, 0, 0})), DateTime.getToday(), "minute")));
+        assertThat(result, is(CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(5), new int[] {2000, 1, 1, 0, 0})), DateTime.getToday(), "minute")));
 
         result = context.resolveExpressionRef("CalculateAgeSeconds").getExpression().evaluate(context);
-        assertThat(result, is((Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(6), new int[] {2000, 1, 1, 0, 0, 0})), DateTime.getToday(), "second")));
+        assertThat(result, is(CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(6), new int[] {2000, 1, 1, 0, 0, 0})), DateTime.getToday(), "second")));
 
         result = context.resolveExpressionRef("CalculateAgeUncertain").getExpression().evaluate(context);
         Integer low = (Integer)CalculateAgeAtEvaluator.calculateAgeAt(new DateTime().withPartial(new Partial(DateTime.getFields(2), new int[] {2000, 12})), DateTime.getToday(), "month");
